Data Science files
Two files of data science for example one source code file and the other is document source code file. This task is about to unsupervised machine learning and in this task, predict find optimum number of clusters for k-means & determine the value of K. For this purpose, use python and Google Colaboratory tool.